station: move new float derived types to core.util
