+++ obspy/station/util.py
@@ -11,6 +11,8 @@ Utility objects.
 """
 from obspy import UTCDateTime
 from obspy.core.util.base import ComparingObject
+from obspy.core.util.obspy_types import FloatWithUncertaintiesAndUnit, \
+    FloatWithUncertaintiesFixedUnit
 import re
 
 
@@ -458,84 +460,6 @@ class Site(ComparingObject):
             town=self.town, county=self.county, region=self.region,
             country=self.country)
         return ret
-
-
-class FloatWithUncertainties(float):
-    """
-    Helper class to inherit from and which stores a float with a given valid
-    range, upper/lower uncertainties and eventual additional attributes.
-    """
-    _minimum = float("-inf")
-    _maximum = float("inf")
-
-    def __new__(cls, value, **kwargs):
-        if not cls._minimum <= float(value) <= cls._maximum:
-            msg = "value %s out of bounds (%s, %s)"
-            msg = msg % (value, cls._minimum, cls._maximum)
-            raise ValueError(msg)
-        return super(FloatWithUncertainties, cls).__new__(cls, value)
-
-    def __init__(self, value, lower_uncertainty=None, upper_uncertainty=None):
-        self.lower_uncertainty = lower_uncertainty
-        self.upper_uncertainty = upper_uncertainty
-
-    @property
-    def unit(self):
-        return self._unit
-
-    def __add__(self, other):
-        return self.__class__(float(self) + other, **self.__dict__)
-
-    def __iadd__(self, other):
-        self = self.__class__(float(self) + other, **self.__dict__)
-
-    def __mul__(self, other):
-        return self.__class__(float(self) * other, **self.__dict__)
-
-    def __imul__(self, other):
-        self = self.__class__(float(self) * other, **self.__dict__)
-
-
-class FloatWithUncertaintiesFixedUnit(FloatWithUncertainties):
-    """
-    Helper class to inherit from, set fixed unit in derived class as class
-    attribute.
-    """
-    _unit = ""
-
-    def __init__(self, value, lower_uncertainty=None, upper_uncertainty=None):
-        super(FloatWithUncertaintiesFixedUnit, self).__init__(
-            value, lower_uncertainty=lower_uncertainty,
-            upper_uncertainty=upper_uncertainty)
-
-    @property
-    def unit(self):
-        return self._unit
-
-    @unit.setter
-    def unit(self, value):
-        msg = "Unit is fixed by StationXML for this object class."
-        raise ValueError(msg)
-
-
-class FloatWithUncertaintiesAndUnit(FloatWithUncertainties):
-    """
-    Helper class to inherit from.
-    """
-    def __init__(self, value, lower_uncertainty=None, upper_uncertainty=None,
-                 unit=None):
-        super(FloatWithUncertaintiesAndUnit, self).__init__(
-            value, lower_uncertainty=lower_uncertainty,
-            upper_uncertainty=upper_uncertainty)
-        self._unit = unit
-
-    @property
-    def unit(self):
-        return self._unit
-
-    @unit.setter
-    def unit(self, value):
-        self._unit = value
 
 
 class Latitude(FloatWithUncertaintiesFixedUnit):
